Update NetworkManager
This is the path for nm-openvpn-service on pop-os: '/usr/lib/NetworkManager/nm-openvpn-service'. All of the executables are under the 'NetworkManager' folder in pop-os, which resulted in log entries in aa-logprof (even though openvpn seemed to function normally for whatever reason). Additionally, the executables did not exist under '/usr/lib' in my filesystem; only under '/usr/lib/NetworkManager'. The respective profiles for each of the executables will need to be changed accordingly.