fix(abs): base: allow inherited mesa cache
mesa 24.2 introduced a shader disk cache which is enabled by default, which opens quite a lot of fd. They are not closed and get inherited by child programs.

Denying it can cause crash, so we are allowing it globally while the issue is beeing fixed in mesa.

See: https://gitlab.freedesktop.org/mesa/mesa/-/issues/11810