feat(profile): Improve restriction of bwrap when used with glycin.
Bwrap needs privileges to create a sandbox. If the sandbox runs in the same
profile as bwrap, then it runs with a lot of access. Most of these accesses are
either dropped early by bwrap or not available from within the sandbox.

It is still a good practice to ensure that bwrap and the sandboxed app run in
different profile (separation, defence in depth...).

However, due to the use of the no-new-pris flag by bwrap, this requires stacking
bwrap & the app profile together. It is not a security issue (on the contrary).
But it may be complex to manage.