Fixes for many profiles
This commit contains fixes many profiles

Rules in this commit were generated by running autopkgtest for all
packages in Ubuntu Questing. Denials were converted to rules thanks to
aa-logprof.

Rules have been generalized manually, but due to the very high amount
of rules, many rules still need to be improved:

Known issues:
 - Some abstractions where added by aa-logprof, but in many cases they
   can be too permissive, and adding individual rules would be preferable
 - There was MANY rules to handle so I only took a small time converting
   and integrating these rules into profiles. All these rules need to be
   carefully reviewed

This commit is only a preliminary result and the final goal is to
integrate such a pipeling into regular package deployment pipelines so
that regressions are quickly detected

Signed-off-by: Maxime Bélair <maxime.belair@canonical.com>