feat(profiles): add global support for glycin loaders
Globally add bwrap transition to the glycin profile in the gtk{-sẗrict} abstractions.

It can be overwritten in profile when bwrap is needed for other purposes.

Only enabled on apparmor 4.1, as older version do not support priority rule, and
are not concerned by this resent update.this resent update.

fix #884
fix #886
fix #887
fix #881